Switched Skills page icons into Boostrap's Puzzle icon
This is still a placeholder, until a nicer-looking one can
be conjured up by our group. The icon was obtained from
https://icons.getbootstrap.com/icons/puzzle/ and is MIT licensed:

The MIT License (MIT)

Copyright (c) 2019-2020 The Bootstrap Authors

Permission is hereby granted, free of charge, to any person obtaining
a copy of this software and associated documentation files (the "Software"),
to deal in the Software without restriction, including without limitation
the rights to use, copy, modify, merge, publish, distribute, sublicense,
and/or sell copies of the Software, and to permit persons to whom the Software
is furnished to do so, subject to the following conditions:

The above copyright notice and this permission notice shall be included in all copies
or substantial portions of the Software.